% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{get_df}
\alias{get_df}
\title{Get chat history as dataframe}
\usage{
get_df(chatdata)
}
\arguments{
\item{chatdata}{Chat data from parse_users_chat_data.}
}
\value{
A dataframe/tibble.
}
\description{
Rows that cannot be parsed are at the top, and rows that parsed successfully are at the bottom.
}
\examples{
csv_file <- file.path(system.file("extdata", package = "chatlogr"), "mtcars.csv")
chatdata <- parse_users_chat_data(csv_file, idcol = "user_id", nrows = 5)
get_df(chatdata)
}
