% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{get_failed_ids}
\alias{get_failed_ids}
\title{Get ids whose data have failed to parse}
\usage{
get_failed_ids(chatdata)
}
\arguments{
\item{chatdata}{Chat data from parse_users_chat_data.}
}
\value{
A vector of user ids.
}
\description{
Get ids whose data have failed to parse
}
\examples{
csv_file <- file.path(system.file("extdata", package = "chatlogr"), "mtcars.csv")
chatdata <- parse_users_chat_data(csv_file, "user_id", nrows = 5)
get_failed_ids(chatdata)
}
