% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{get_success_ids}
\alias{get_success_ids}
\title{Get ids whose data have been parsed sucessfully}
\usage{
get_success_ids(chatdata)
}
\arguments{
\item{chatdata}{Chat data from parse_users_chat_data.}
}
\value{
A vector of user ids.
}
\description{
Get ids whose data have been parsed sucessfully
}
\examples{
csv_file <- file.path(system.file("extdata", package = "chatlogr"), "mtcars.csv")
chatdata <- parse_users_chat_data(csv_file, idcol = "user_id")
get_success_ids(chatdata)
}
