% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{parse_user_chat_data}
\alias{parse_user_chat_data}
\title{Parse a single user's raw chat data}
\usage{
parse_user_chat_data(
  user_chat_data,
  user_id,
  join_str = "\\"\\"",
  verbose = TRUE
)
}
\arguments{
\item{user_chat_data}{Raw chat data for a single user. Must only include chat data columns.}

\item{user_id}{User ID in character.}

\item{join_str}{Join string in character. Defaults to '\"\"'.}

\item{verbose}{Logical. Defaults to TRUE.}
}
\value{
JSON data as list.
}
\description{
Parse a single user's raw chat data
}
\examples{
csv_file <- file.path(system.file("extdata", package = "chatlogr"), "mtcars.csv")
user_chat_data <- read_relevant_data_columns(csv_file, "user_id", "chathistory")[1, ]
user_chatdata <- parse_user_chat_data(user_chat_data, "vs", join_str = '\"\"')
}
