% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{parse_users_chat_data}
\alias{parse_users_chat_data}
\title{Parse all user data}
\usage{
parse_users_chat_data(
  csv_file,
  idcol = "ResponseId",
  chat_col_patterns = c("chathistory"),
  chat_cols = "",
  join_str = "\\"\\"",
  nrows = Inf,
  verbose = FALSE
)
}
\arguments{
\item{csv_file}{Path to csv file in character.}

\item{idcol}{Name of column that is the user ID column in character. Must contain unique values.}

\item{chat_col_patterns}{Patterns to match chat columns in character. Defaults to "chathistory".}

\item{chat_cols}{Chat columns in character. Defaults "".}

\item{join_str}{Join string in character. Defaults to '\"\"'.}

\item{nrows}{Number of rows to process. Defaults to Inf. Use fewer rows to debug the function/parser.}

\item{verbose}{Whether to print processing status. Defaults to FALSE.}
}
\value{
List with info, info_df, json_list, df_success, df_fail keys.
}
\description{
Parse all user data
}
\examples{
csv_file <- file.path(system.file("extdata", package = "chatlogr"), "mtcars.csv")
chatdata <- parse_users_chat_data(csv_file, "user_id")
}
