% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{read_relevant_data_columns}
\alias{read_relevant_data_columns}
\title{Read csv file to get matching data columns}
\usage{
read_relevant_data_columns(
  csv_file = NULL,
  dat = NULL,
  idcol = NULL,
  chat_col_patterns = c("chathistory"),
  chat_cols = ""
)
}
\arguments{
\item{csv_file}{Path to csv file in character. Defaults to NULL.}

\item{dat}{Dataframe/tibble. Defaults to NULL.}

\item{idcol}{ID column in character. Defaults to NULL.}

\item{chat_col_patterns}{Patterns to match chat columns in character. Defaults to "chathistory".}

\item{chat_cols}{Chat columns in character. Defaults to "".}
}
\value{
Dataframe/tibble with matching columns.
}
\description{
Read csv file to get matching data columns
}
\examples{
csv_file <- file.path(system.file("extdata", package = "chatlogr"), "mtcars.csv")
dat <- read_relevant_data_columns(csv_file, idcol = "vs", chat_col_patterns = "mpg")
}
