% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{write_to_csv}
\alias{write_to_csv}
\title{Write chat history to csv file}
\usage{
write_to_csv(chatdata, output_file)
}
\arguments{
\item{chatdata}{Chat data from parse_users_chat_data.}

\item{output_file}{Output file name. Must end with '.csv'.}
}
\value{
TRUE if successful
}
\description{
Write only chat history to csv file. Does not include other information contained in the full chat data. To save the full chat data, use `write_to_json()`.
}
\examples{
\dontrun{
write_to_csv(chatdata, "chatdata.csv")
}
}
