% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{write_to_json}
\alias{write_to_json}
\title{Write chat data to json file}
\usage{
write_to_json(chatdata, output_file)
}
\arguments{
\item{chatdata}{Chat data from parse_users_chat_data.}

\item{output_file}{Output file name. Must end with '.json'.}
}
\value{
TRUE if successful
}
\description{
Write chat data to json file
}
\examples{
\dontrun{
write_to_json(chatdata, "chatdata.json")
}
}
