% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{get_user_chat}
\alias{get_user_chat}
\title{Get and view chat history of a user}
\usage{
get_user_chat(chatdata, user_id = NULL, random_user = FALSE)
}
\arguments{
\item{chatdata}{Chat data from parse_users_chat_data.}

\item{user_id}{A user ID in character. Default is NULL.}

\item{random_user}{A logical value. If TRUE, a random user will be selected. Default is FALSE.}
}
\value{
Chat history of a user in a tibble.
}
\description{
Get and view chat history of a user
}
\examples{
csv_file <- file.path(system.file("extdata", package = "chatlogr"), "mtcars.csv")
chatdata <- parse_users_chat_data(csv_file, "user_id")
get_user_chat(chatdata, user_id = 1)
}
