% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlogr.R
\name{parse_users_chat_data}
\alias{parse_users_chat_data}
\title{Parse all user data}
\usage{
parse_users_chat_data(
  csv_file = NULL,
  dat = NULL,
  idcol = "ResponseId",
  chat_col_patterns = c("chathistory"),
  chat_cols = "",
  join_str = "\\"\\"",
  nrows = Inf,
  verbose = FALSE
)
}
\arguments{
\item{csv_file}{Path to csv file in character. Defaults to NULL.}

\item{dat}{Dataframe/tibble with raw chat data in one or more columns. Defaults to NULL.}

\item{idcol}{Name of column that is the user ID column in character. Must contain unique values.}

\item{chat_col_patterns}{Patterns to match chat columns in character. Defaults to "chathistory".}

\item{chat_cols}{Chat columns in character. Defaults "".}

\item{join_str}{Join string in character. Defaults to '\"\"'.}

\item{nrows}{Number of rows to process. Defaults to Inf. Use fewer rows to debug the function/parser.}

\item{verbose}{Whether to print processing status. Defaults to FALSE.}
}
\value{
A list with the following keys:
\describe{
  \item{info}{A vector indicating total number of rows processed, number of success/error/warning parses.}
  \item{info_df}{A dataframe/tibble with info on each user's chat data, where the status column indicates parsing status and mobile column indicates mobile (1) or not (0).}
  \item{json_list}{A list of JSON data for each user. Contains full data and can be a very large list.}
  \item{df_success}{A dataframe/tibble with successful parses. Contains only chat messages and not other data.}
  \item{df_fail}{A dataframe/tibble with failed parses. Contains only chat messages and not other data.}
}
}
\description{
Parse all user data
}
\examples{
csv_file <- file.path(system.file("extdata", package = "chatlogr"), "mtcars.csv")
chatdata <- parse_users_chat_data(csv_file, idcol = "user_id")
}
